refactor: Move all painters into own folder
